Refactoring the context support: moved into the base class EntityRecognizer
