+++ presidio-analyzer/tests/test_analyzer_engine.py
@@ -19,48 +19,6 @@ from analyzer.nlp_engine import SpacyNlpEngine, NlpArtifacts
 from analyzer.predefined_recognizers import IpRecognizer, UsSsnRecognizer
 from tests.mocks import MockNlpEngine
 
-@pytest.fixture(scope="class")
-def sentences_with_context(request):
-    path = os.path.dirname(__file__) + '/data/context_sentences_tests.txt'
-    f = open(path, "r")
-    if not f.mode == 'r':
-        return []
-    content = f.read()
-    f.close()
-    lines = content.split('\n')
-    # remove empty lines
-    lines = list(filter(lambda k: k.strip(), lines))
-    # remove comments
-    lines = list(filter(lambda k: k[0] != '#', lines))
-    test_items = []
-    for i in range(len(lines)):
-        if i % 2 == 1:
-            continue
-        recognizer = None
-        entity_type = lines[i].strip()
-        if entity_type == "IP_ADDRESS":
-            recognizer = ip_recognizer
-        elif entity_type == "US_SSN":
-            recognizer = us_ssn_recognizer
-        elif entity_type == "PHONE_NUMBER":
-            recognizer = phone_recognizer
-        elif entity_type == "US_ITIN":
-            recognizer = us_itin_recognizer
-        elif entity_type == "US_DRIVER_LICENSE":
-            recognizer = us_license_recognizer
-        elif entity_type == "US_BANK_NUMBER":
-            recognizer = us_bank_recognizer
-        elif entity_type == "US_PASSPORT":
-            recognizer = us_passport_recognizer
-        else:
-            # will fail the test in its turn
-            print("bad type: ", entity_type)
-            return []
-        test_items.append((lines[i+1].strip(),
-                           recognizer,
-                           [lines[i].strip()]))
-    request.cls.context_sentences = test_items
-
 class RecognizerStoreApiMock(RecognizerStoreApi):
     """
     A mock that acts as a recognizers store, allows to add and get recognizers
@@ -127,7 +85,6 @@ us_license_recognizer = UsLicenseRecognizer()
 us_bank_recognizer = UsBankRecognizer()
 us_passport_recognizer = UsPassportRecognizer()
 
-@pytest.mark.usefixtures("sentences_with_context")
 class TestAnalyzerEngine(TestCase):
 
     def __init__(self, *args, **kwargs):
@@ -325,28 +282,4 @@ class TestAnalyzerEngine(TestCase):
         new_field.minScore = '0.5'
         with pytest.raises(ValueError):
             analyze_engine.Apply(request, None)
-
-    # Context tests
-    def test_text_with_context_improves_score(self):
-        # Currently we have 19 sentences, this is a sanity
-        assert(len(self.context_sentences) == 19)
-        analyze_engine = AnalyzerEngine()
-        nlp_engine = SpacyNlpEngine()
-        
-        for item in self.context_sentences:
-            text = item[0]
-            recognizer = item[1]
-            entities = item[2]
-            nlp_artifacts = nlp_engine.process_text(text, "en")
-            results = recognizer.analyze(text, entities)
-
-            enhanced_results = analyze_engine._AnalyzerEngine__enhance_using_context(text,
-                                                                                     recognizer,
-                                                                                     results,
-                                                                                     nlp_artifacts)
-            assert(results)
-            assert(len(results) == len(enhanced_results))
-            for i in range(len(results)):
-                assert(results[i].score < enhanced_results[i].score)
-
